Move ticklist logic to new file
