+++ application.py
@@ -13,6 +13,9 @@ from werkzeug.urls import url_parse
 
 import db.firebase_controller as firebase_controller
 from config import *
+from utils.utils import *
+
+import ticklist_handler
 
 # create the application object
 app = Flask(__name__)
@@ -73,12 +76,13 @@ def get_path_from_gym_name(gym_name):
 
 def get_wall_radius(wall_path=None):
     """
-    wall path is expected to be: 'gym/wall'
+    Wall path is expected to be: 'gym/wall'
     """
     if session.get('walls_radius', ''):
         return session['walls_radius'][wall_path]
     else:
         return firebase_controller.get_walls_radius_all()[wall_path]
+
 
 def get_stats():
     """
@@ -103,16 +107,6 @@ def get_stats():
         'Routes': total_routes,
         'Gyms': total_gyms
     }
-
-def load_boulder_from_request(request):
-    """
-    Replace boulder data from valid Python to valid JS
-    """
-    return json.loads(
-                request.form.get('boulder_data')
-                .replace('\'', '"')
-                .replace('True', 'true')
-                .replace('False', 'false'))
 
 
 @app.route('/', methods=['GET', 'POST'])
@@ -408,40 +402,8 @@ def logout():
 @login_required
 def tick_list():
     if request.method == 'POST':
-        # needed values: gym, id, section, is_done
-        boulder = {
-            'gym': request.form.get('gym'),
-            'iden': list(
-                firebase_controller.get_boulder_by_name(
-                    request.form.get('gym'), 
-                    request.form.get('name')
-                ).keys()
-            )[0],
-            'is_done': True if request.form.get('is_done', '') else False,
-            'section': request.form.get('section')
-        }
-        # update user's ticklist
-        current_user.ticklist = [
-            TickListProblem(p) for p in firebase_controller.put_boulder_in_ticklist(boulder, current_user.id)
-        ]
-    # get boulders in ticklist and extra required values
-    boulder_list = [
-        firebase_controller.get_ticklist_boulder(problem) for problem in current_user.ticklist
-    ]
-    unique_sections = dict()
-    walls_list = []
-    for boulder in boulder_list:
-        boulder['feet'] = FEET_MAPPINGS[boulder['feet']]
-        boulder['safe_name'] = secure_filename(boulder['name'])
-        boulder['radius'] = get_wall_radius(boulder['gym']+ '/' + boulder['section'])
-        boulder['color'] = BOULDER_COLOR_MAP[boulder['difficulty']]
-        if boulder['gym'] not in unique_sections.keys() and boulder['section'] not in unique_sections.values():
-            unique_sections[boulder['gym']] = boulder['section']
-            walls_list.append({
-                'gym_name': firebase_controller.get_gym_pretty_name(boulder['gym']),
-                'image': boulder['section'], 
-                'name': firebase_controller.get_wall_name(boulder['gym'], boulder['section'])
-            })
+        current_user.ticklist = ticklist_handler.add_boulder_to_ticklist(request, current_user)
+    boulder_list, walls_list = ticklist_handler.load_user_ticklist(current_user)
     return render_template(
         'tick_list.html', 
         boulder_list = boulder_list,
@@ -452,23 +414,7 @@ def tick_list():
 @app.route('/delete_ticklist_problem', methods=['POST'])
 def delete_ticklist_problem():
     if request.method == 'POST':
-        # needed values: gym, id, section, is_done
-        boulder_data = load_boulder_from_request(request)
-        boulder = {
-            'gym': boulder_data.get('gym'),
-            'iden': list(
-                firebase_controller.get_boulder_by_name(
-                    boulder_data.get('gym'), 
-                    request.form.get('name')
-                ).keys()
-            )[0],
-            's_done': boulder_data.get('is_done'),
-            'section': boulder_data.get('section')
-        }
-        # update user's ticklist
-        current_user.ticklist = [
-            TickListProblem(p) for p in firebase_controller.delete_boulder_in_ticklist(boulder, current_user.id)
-        ]
+        current_user.ticklist = ticklist_handler.delete_problem_from_ticklist(request, current_user)
     return redirect(url_for('tick_list'))
